requirements.txt
selenium
resend
python-dotenv